Mark package as ESM with type: module
