+++ .changeset/provide-esm-cjs-builds.md
@@ -2,4 +2,4 @@
 "eslint-plugin-es-x": major
 ---
 
-Build publish output with tsdown and provide both ESM and CommonJS package entry points.
+Build publish output with tsdown, mark the package as ESM with `"type": "module"`, and provide both ESM and CommonJS package entry points.
