+++ eslint-internal/config/_browser-globals.cjs
