+++ eslint-internal/rules/es-x-rule/require-allow-tested-property-schema.cjs
@@ -1,11 +1,15 @@
 "use strict"
 
-const { ReferenceTracker, READ } = require("@eslint-community/eslint-utils")
+const {
+    ReferenceTracker,
+    READ,
+    ESM,
+} = require("@eslint-community/eslint-utils")
 const path = require("node:path")
 const {
     defineSchemaChecker,
     getProperty,
-} = require("../../utils/schema-checker")
+} = require("../../utils/schema-checker.cjs")
 
 /**
  * @typedef {import("estree").Node} Node
@@ -35,20 +39,28 @@ module.exports = {
 
         const traceMap = Object.fromEntries(
             Object.entries({
-                "lib/util/define-static-properties-handler": {
+                "lib/util/define-static-properties-handler/index.ts": {
                     defineStaticPropertiesHandler: { [READ]: true },
+                    [ESM]: true,
                 },
-                "lib/util/define-prototype-properties-handler": {
+                "lib/util/define-prototype-properties-handler/index.ts": {
                     definePrototypePropertiesHandler: { [READ]: true },
+                    [ESM]: true,
                 },
-                "lib/util/define-nonstandard-prototype-properties-handler": {
-                    defineNonstandardPrototypePropertiesHandler: {
-                        [READ]: true,
+                "lib/util/define-nonstandard-prototype-properties-handler/index.ts":
+                    {
+                        defineNonstandardPrototypePropertiesHandler: {
+                            [READ]: true,
+                        },
+                        [ESM]: true,
                     },
-                },
-                "lib/util/define-nonstandard-static-properties-handler": {
-                    defineNonstandardStaticPropertiesHandler: { [READ]: true },
-                },
+                "lib/util/define-nonstandard-static-properties-handler/index.ts":
+                    {
+                        defineNonstandardStaticPropertiesHandler: {
+                            [READ]: true,
+                        },
+                        [ESM]: true,
+                    },
             }).map(([filePath, properties]) => {
                 const absolutePath = path.join(__dirname, "../../..", filePath)
                 const relativePath = path.relative(
@@ -59,7 +71,7 @@ module.exports = {
             }),
         )
 
-        if (tracker.iterateCjsReferences(traceMap).next().done) {
+        if (tracker.iterateEsmReferences(traceMap).next().done) {
             return {}
         }
