+++ eslint-internal/rules/my-block-scoped-var.cjs
