+++ eslint-internal/rules/my-no-instanceof-array.cjs
