+++ eslint-internal/rules/my-no-instanceof-wrapper.cjs
