+++ eslint-internal/rules/my-no-literal-call.cjs
