+++ eslint-internal/rules/my-no-this-in-static.cjs
