+++ eslint-internal/rules/my-no-useless-rest-spread.cjs
