+++ eslint-internal/rules/my-prefer-for-of.cjs
