+++ eslint-internal/utils/index.cjs
